+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 # v2022.5.1-beta
 - wrapper - add jslint-addon for codemirror
-- allow jslint.mjs to auto-export itself to globalThis when given search-param `?jslint_export_global=1`
+- allow jslint.mjs to auto-export itself to globalThis when given search-param `?window_jslint=1`
 - wrapper - add jslint-extension for vscode
 - bugfix - fix jslint falsely believing megastring literals `0` and `1` are similar
 - bugfix - fix function jstestOnExit() from exiting prematurely and suppressing additional error-messages